photonix-backend/face-detection-service/requirements.txt
# Face Detection & Recognition (Simple ARM/M1 Mac compatible)
insightface==0.7.3
onnxruntime==1.16.3

# Image Processing
opencv-python-headless==4.8.1.78
Pillow==10.1.0
numpy==1.24.3

# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Utilities
python-multipart==0.0.6
requests==2.31.0
scikit-learn==1.3.2
